+++ lib/jmap/core/capability/calendar_event_capability.dart
@@ -6,8 +6,14 @@ part 'calendar_event_capability.g.dart';
 @JsonSerializable(includeIfNull: false)
 class CalendarEventCapability extends CapabilityProperties {
   final List<String>? replySupportedLanguage;
+  final bool? supportFreeBusyQuery;
+  final bool? counterSupport;
 
-  CalendarEventCapability({this.replySupportedLanguage});
+  CalendarEventCapability({
+    this.replySupportedLanguage,
+    this.supportFreeBusyQuery,
+    this.counterSupport,
+  });
   
   factory CalendarEventCapability.fromJson(Map<String, dynamic> json) 
     => _$CalendarEventCapabilityFromJson(json);
@@ -19,5 +25,9 @@ class CalendarEventCapability extends CapabilityProperties {
   }
   
   @override
-  List<Object?> get props => [replySupportedLanguage];
+  List<Object?> get props => [
+    replySupportedLanguage,
+    supportFreeBusyQuery,
+    counterSupport,
+  ];
 }