+++ lib/jmap/core/capability/calendar_event_capability.g.dart
@@ -12,6 +12,8 @@ CalendarEventCapability _$CalendarEventCapabilityFromJson(
       replySupportedLanguage: (json['replySupportedLanguage'] as List<dynamic>?)
           ?.map((e) => e as String)
           .toList(),
+      supportFreeBusyQuery: json['supportFreeBusyQuery'] as bool?,
+      counterSupport: json['counterSupport'] as bool?,
     );
 
 Map<String, dynamic> _$CalendarEventCapabilityToJson(
@@ -25,5 +27,7 @@ Map<String, dynamic> _$CalendarEventCapabilityToJson(
   }
 
   writeNotNull('replySupportedLanguage', instance.replySupportedLanguage);
+  writeNotNull('supportFreeBusyQuery', instance.supportFreeBusyQuery);
+  writeNotNull('counterSupport', instance.counterSupport);
   return val;
 }