+++ lib/jmap/core/method/request/set_method.dart
@@ -2,16 +2,18 @@ import 'package:jmap_dart_client/jmap/account_id.dart';
 import 'package:jmap_dart_client/jmap/core/id.dart';
 import 'package:jmap_dart_client/jmap/core/method/method.dart';
 import 'package:jmap_dart_client/jmap/core/patch_object.dart';
+import 'package:jmap_dart_client/jmap/core/request/result_reference.dart';
 import 'package:jmap_dart_client/jmap/core/state.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 abstract class SetMethod<T> extends MethodRequiringAccountId
-    with OptionalIfInState, OptionalCreate<T>, OptionalDestroy, OptionalUpdate {
+    with OptionalIfInState, OptionalCreate<T>, OptionalDestroy, OptionalUpdate,
+    OptionalReferenceDestroy {
   SetMethod(AccountId accountId) : super(accountId);
 }
 
 abstract class SetMethodNoNeedAccountId<T> extends Method
-    with OptionalCreate<T>, OptionalDestroy, OptionalUpdate {
+    with OptionalCreate<T>, OptionalDestroy, OptionalUpdate, OptionalReferenceDestroy {
   SetMethodNoNeedAccountId() : super();
 }
 
@@ -20,7 +22,7 @@ mixin OptionalIfInState {
   State? ifInState;
 
   void addIfInState(State? state) {
-    this.ifInState = state;
+    ifInState = state;
   }
 }
 
@@ -29,16 +31,12 @@ mixin OptionalCreate<T> {
   Map<Id, T>? create;
 
   void addCreates(Map<Id, T> creates) {
-    if (create == null) {
-      create = Map<Id, T>();
-    }
+    create ??= <Id, T>{};
     create?.addAll(creates);
   }
 
   void addCreate(Id id, T createItem) {
-    if (create == null) {
-      create = Map<Id, T>();
-    }
+    create ??= <Id, T>{};
     create?.addAll({id: createItem});
   }
 }
@@ -48,9 +46,7 @@ mixin OptionalUpdate {
   Map<Id, PatchObject>? update;
 
   void addUpdates(Map<Id, PatchObject> updates) {
-    if (update == null) {
-      update = Map<Id, PatchObject>();
-    }
+    update ??= <Id, PatchObject>{};
     update?.addAll(updates);
   }
 }
@@ -60,9 +56,7 @@ mixin OptionalDestroy {
   Set<Id>? destroy;
 
   void addDestroy(Set<Id> values) {
-    if (destroy == null) {
-      destroy = Set();
-    }
+    destroy ??= <Id>{};
     destroy?.addAll(values);
   }
 }
@@ -72,9 +66,16 @@ mixin OptionalUpdateSingleton<T> {
   Map<Id, T>? updateSingleton;
 
   void addUpdatesSingleton(Map<Id, T> updates) {
-    if (updateSingleton == null) {
-      updateSingleton = Map<Id, T>();
-    }
+    updateSingleton ??= <Id, T>{};
     updateSingleton?.addAll(updates);
+  }
+}
+
+mixin OptionalReferenceDestroy {
+  @JsonKey(name: '#destroy', includeIfNull: false)
+  ResultReference? referenceDestroy;
+
+  void addReferenceDestroy(ResultReference resultReferenceDestroy) {
+    referenceDestroy = resultReferenceDestroy;
   }
 }