+++ test/jmap/email/changes/get_changes_email_test.dart
@@ -218,7 +218,6 @@ void main() {
         },
         headers: {
           "accept": "application/json;jmapVersion=rfc-8621",
-          "content-length": 1733
         }
       );