+++ test/jmap/email/get/get_email_method_test.dart
@@ -112,7 +112,6 @@ void main() {
         },
         headers: {
           "accept": "application/json;jmapVersion=rfc-8621",
-          "content-length": 1018
         }
       );