+++ test/jmap/email/get_list_email_test.dart
@@ -226,7 +226,6 @@ void main() {
           },
           headers: {
             "accept": "application/json;jmapVersion=rfc-8621",
-            "content-length": 1187
           }
       );