+++ test/jmap/email/sort_list_email_test.dart
@@ -238,7 +238,6 @@ void main() {
         },
         headers: {
           "accept": "application/json;jmapVersion=rfc-8621",
-          "content-length": 1234
         }
     );